Conclusão do trabalho1

+++ proc.h
@@ -50,7 +50,7 @@ struct proc {
   struct inode *cwd;           // Current directory
   char name[16];               // Process name (debugging)
   int tickets;
-  int qtd;
+  long int qtd;
 };
 
 // Process memory is laid out contiguously, low addresses first:
